# apps/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: AgentKern API
  description: |
    AgentKern - The Kernel for Verified AI Agents
    
    This API provides access to all AgentKern pillars:
    - Identity: Agent registration and signing
    - Gate: Policy verification and guardrails
    - Synapse: State and memory management
    - Arbiter: Coordination and locking
    - Sovereign: Data residency compliance
    - Treasury: Agent-to-agent payments
  version: 1.0.0
  contact:
    name: AgentKern Team
    url: https://agentkern.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.agentkern.dev/v1
    description: Production

tags:
  - name: Identity
    description: Agent identity and signing (Passport)
  - name: Gate
    description: Policy verification and guardrails
  - name: Synapse
    description: State and memory management
  - name: Arbiter
    description: Coordination and locking
  - name: Sovereign
    description: Data residency compliance
  - name: Treasury
    description: Agent-to-agent payments

paths:
  # ============================================================================
  # IDENTITY
  # ============================================================================
  /identity/register:
    post:
      operationId: register_agent
      summary: Register a new agent
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Agent registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentIdentity'
        '400':
          $ref: '#/components/responses/BadRequest'

  /identity/{agentId}:
    get:
      operationId: get_identity
      summary: Get agent identity
      tags: [Identity]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentIdentity'
        '404':
          $ref: '#/components/responses/NotFound'

  /identity/{agentId}/sign:
    post:
      operationId: sign_action
      summary: Sign an action to create liability proof
      tags: [Identity]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignRequest'
      responses:
        '200':
          description: Liability proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiabilityProof'

  /identity/verify:
    post:
      operationId: verify_proof
      summary: Verify a liability proof signature
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LiabilityProof'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyProofResponse'

  /identity/{agentId}/trust-score:
    get:
      operationId: get_trust_score
      summary: Get agent's trust score
      tags: [Identity]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Trust score
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: number
                    minimum: 0
                    maximum: 100

  # ============================================================================
  # GATE
  # ============================================================================
  /gate/verify:
    post:
      operationId: verify_action
      summary: Verify if an action is allowed by policies
      tags: [Gate]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResult'

  /gate/attest:
    post:
      operationId: generate_attestation
      summary: Generate hardware-level TEE attestation
      tags: [Gate]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nonce]
              properties:
                nonce:
                  type: string
                  description: User-supplied nonce for the attestation quote
      responses:
        '200':
          description: TEE Attestation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attestation'

  /gate/carbon/budget/{agentId}:
    get:
      operationId: get_carbon_budget
      summary: Get agent's carbon budget
      tags: [Gate]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Carbon budget
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarbonBudget'
    post:
      operationId: set_carbon_budget
      summary: Set agent's carbon budget
      tags: [Gate]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarbonBudget'
      responses:
        '200':
          description: Budget updated

  /gate/carbon/usage/{agentId}:
    get:
      operationId: get_carbon_usage
      summary: Get agent's carbon usage summary
      tags: [Gate]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Carbon usage summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarbonUsage'

  /gate/policies:
    get:
      operationId: list_policies
      summary: List all active policies
      tags: [Gate]
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Policy'
    post:
      operationId: create_policy
      summary: Create a new policy
      tags: [Gate]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
      responses:
        '201':
          description: Policy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'

  # ============================================================================
  # SYNAPSE
  # ============================================================================
  /synapse/state/{agentId}:
    get:
      operationId: get_state
      summary: Get agent state
      tags: [Synapse]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentState'
    put:
      operationId: set_state
      summary: Update agent state (CRDT merge)
      tags: [Synapse]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Updated state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentState'

  /synapse/intent/{agentId}:
    post:
      operationId: start_intent
      summary: Start a new intent path
      tags: [Synapse]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntentRequest'
      responses:
        '201':
          description: Intent path started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntentPath'

  /synapse/drift/{agentId}:
    get:
      operationId: check_drift
      summary: Check if agent has drifted from intent
      tags: [Synapse]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Drift check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftResult'

  # ============================================================================
  # ARBITER
  # ============================================================================
  /arbiter/coordinate:
    post:
      operationId: request_coordination
      summary: Request coordination for a resource
      tags: [Arbiter]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoordinationRequest'
      responses:
        '200':
          description: Coordination result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoordinationResult'

  /arbiter/lock:
    post:
      operationId: acquire_lock
      summary: Acquire a lock on a resource
      tags: [Arbiter]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LockRequest'
      responses:
        '200':
          description: Lock acquired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessLock'
        '409':
          description: Resource already locked
    delete:
      operationId: release_lock
      summary: Release a lock
      tags: [Arbiter]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId, resource]
              properties:
                agentId:
                  type: string
                resource:
                  type: string
      responses:
        '200':
          description: Lock released

  # ============================================================================
  # SOVEREIGN
  # ============================================================================
  /sovereign/can-transfer:
    post:
      operationId: can_transfer
      summary: Check if data can be transferred between regions
      tags: [Sovereign]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferCheck'
      responses:
        '200':
          description: Transfer allowed
          content:
            application/json:
              schema:
                type: object
                properties:
                  allowed:
                    type: boolean
                  reason:
                    type: string

  /sovereign/compliance:
    post:
      operationId: validate_compliance
      summary: Validate operation compliance
      tags: [Sovereign]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceCheck'
      responses:
        '200':
          description: Compliance result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceResult'

  # ============================================================================
  # TREASURY
  # ============================================================================
  /treasury/balance/{agentId}:
    get:
      operationId: get_balance
      summary: Get agent balance
      tags: [Treasury]
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'

  /treasury/transfer:
    post:
      operationId: transfer
      summary: Transfer funds between agents
      tags: [Treasury]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Transfer result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResult'

  /treasury/can-spend:
    post:
      operationId: can_spend
      summary: Pre-flight check if agent can spend
      tags: [Treasury]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId, amount]
              properties:
                agentId:
                  type: string
                amount:
                  type: number
      responses:
        '200':
          description: Can spend result
          content:
            application/json:
              schema:
                type: object
                properties:
                  allowed:
                    type: boolean
                  remaining:
                    type: number

components:
  parameters:
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    # Common
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string

    # Identity
    RegisterRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        capabilities:
          type: array
          items:
            type: string

    AgentIdentity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        publicKey:
          type: string
        capabilities:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        trustScore:
          type: number

    SignRequest:
      type: object
      required: [action, payload]
      properties:
        action:
          type: string
        payload:
          type: object

    LiabilityProof:
      type: object
      properties:
        action:
          type: string
        agentId:
          type: string
        timestamp:
          type: string
          format: date-time
        signature:
          type: string
        payloadHash:
          type: string

    VerifyProofResponse:
      type: object
      properties:
        valid:
          type: boolean
        reason:
          type: string

    # Gate
    VerifyRequest:
      type: object
      required: [agentId, action]
      properties:
        agentId:
          type: string
        action:
          type: string
        context:
          type: object

    VerificationResult:
      type: object
      properties:
        allowed:
          type: boolean
        reason:
          type: string
        policies:
          type: array
          items:
            type: string

    Policy:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        rules:
          type: array
          items:
            type: object
        enabled:
          type: boolean

    Attestation:
      type: object
      properties:
        platform:
          type: string
          enum: [intel_tdx, amd_sev_snp, simulated]
        quote:
          type: string
          description: Base64 encoded hardware quote
        measurement:
          type: string
          description: Hex encoded enclave measurement
        nonce:
          type: string
        timestamp:
          type: integer

    CarbonBudget:
      type: object
      required: [dailyLimitGrams]
      properties:
        dailyLimitGrams:
          type: number
        monthlyLimitGrams:
          type: number
        blockOnExceed:
          type: boolean

    CarbonUsage:
      type: object
      properties:
        totalCo2Grams:
          type: number
        totalEnergyKwh:
          type: number
        actionCount:
          type: integer
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time

    # Synapse
    AgentState:
      type: object
      properties:
        agentId:
          type: string
        data:
          type: object
        version:
          type: integer
        lastUpdated:
          type: string
          format: date-time

    IntentRequest:
      type: object
      required: [intent, expectedSteps]
      properties:
        intent:
          type: string
        expectedSteps:
          type: integer

    IntentPath:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        intent:
          type: string
        steps:
          type: array
          items:
            type: object
        completedSteps:
          type: integer
        driftScore:
          type: number

    DriftResult:
      type: object
      properties:
        drifted:
          type: boolean
        score:
          type: number

    # Arbiter
    CoordinationRequest:
      type: object
      required: [agentId, resource]
      properties:
        agentId:
          type: string
        resource:
          type: string
        action:
          type: string
        priority:
          type: integer

    CoordinationResult:
      type: object
      properties:
        granted:
          type: boolean
        queuePosition:
          type: integer
        estimatedWait:
          type: integer

    LockRequest:
      type: object
      required: [agentId, resource]
      properties:
        agentId:
          type: string
        resource:
          type: string
        priority:
          type: integer
        ttlSeconds:
          type: integer

    BusinessLock:
      type: object
      properties:
        id:
          type: string
        resource:
          type: string
        holder:
          type: string
        acquiredAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time

    # Sovereign
    TransferCheck:
      type: object
      required: [fromRegion, toRegion, dataType]
      properties:
        fromRegion:
          type: string
        toRegion:
          type: string
        dataType:
          type: string

    ComplianceCheck:
      type: object
      required: [operation, jurisdiction]
      properties:
        operation:
          type: string
        data:
          type: object
        jurisdiction:
          type: string

    ComplianceResult:
      type: object
      properties:
        compliant:
          type: boolean
        violations:
          type: array
          items:
            type: string

    # Treasury
    Balance:
      type: object
      properties:
        balance:
          type: number
        currency:
          type: string
        pending:
          type: number

    TransferRequest:
      type: object
      required: [from, to, amount]
      properties:
        from:
          type: string
        to:
          type: string
        amount:
          type: number
        reference:
          type: string

    TransferResult:
      type: object
      properties:
        transactionId:
          type: string
        status:
          type: string
          enum: [completed, pending, failed]